Release 1.3.2

+++ src/main/java/me/dreamvoid/miraimc/bukkit/MiraiAutoLogin.java
@@ -76,7 +76,7 @@ public class MiraiAutoLogin {
             for(Map<?,?> map : loadAutoLoginList()){
                 Map<?,?> password = (Map<?, ?>) map.get("password");
                 Map<?,?> configuration = (Map<?, ?>) map.get("configuration");
-                Integer Account = (Integer) map.get("account");
+                long Account = (Long) map.get("account");
                 String Password = password.get("value").toString();
                 BotConfiguration.MiraiProtocol Protocol = BotConfiguration.MiraiProtocol.valueOf(configuration.get("protocol").toString());
